Kis 26 crear conexiones memoria cpu

+++ Shared/utils/instrucciones.c
@@ -1,7 +1,7 @@
 #include "instrucciones.h"
 
 
-int definirCodigo(char* id) {
+uint8_t definirCodigo(char* id) {
 	if (string_contains(id, "NO_OP")) {
 		return 1;
 	}
@@ -24,7 +24,7 @@ int definirCodigo(char* id) {
 	}
 }
 
-int getCantidadParametros(int id) {
+int getCantidadParametros(uint8_t id) {
 	switch (id) {
 	case 1: case 2: case 5:
 		return 1;
